changed location; moved the last remaining file up one directory to the hardwareProfiles dir.  also deleted the last of the migration files
