--- frontend/src/pages/hardwareProfiles/migration/types.ts
@@ -1,6 +1,0 @@
-export type ContainerSizeLimits = {
-  minCpu: string | number;
-  maxCpu?: string | number;
-  minMemory: string;
-  maxMemory?: string;
-};
