+++ frontend/src/pages/hardwareProfiles/useHardwareProfilesByFeatureVisibility.ts
