+++ frontend/src/pages/pipelines/global/modelCustomization/trainingHardwareSection/TrainingHardwareProfileFormSection.tsx
@@ -16,7 +16,7 @@ import { HardwareProfileConfig } from '#~/concepts/hardwareProfiles/useHardwareP
 import HardwareProfileCustomize from '#~/concepts/hardwareProfiles/HardwareProfileCustomize';
 import HardwareProfileSelect from '#~/concepts/hardwareProfiles/HardwareProfileSelect';
 import { filterHardwareProfilesForTraining } from '#~/pages/pipelines/global/modelCustomization/utils';
-import { useHardwareProfilesByFeatureVisibility } from '#~/pages/hardwareProfiles/migration/useHardwareProfilesByFeatureVisibility';
+import { useHardwareProfilesByFeatureVisibility } from '#~/pages/hardwareProfiles/useHardwareProfilesByFeatureVisibility.ts';
 import DashboardPopupIconButton from '#~/concepts/dashboard/DashboardPopupIconButton';
 import { SupportedArea, useIsAreaAvailable } from '#~/concepts/areas';
